Moves middlewares to own file
