+++ lib/resources/auth.js
@@ -2,52 +2,7 @@
 
 const express = require('express');
 const HttpStatus = require('http-status-codes');
-
-const utils = require('../utils');
-
-function checkCredentials(creds) {
-  if (!(typeof creds === 'object' && typeof creds.username === 'string' && typeof creds.password === 'string')) {
-    throw new Error('Credentials not provided');
-  }
-}
-
-function requireCredentialsMiddleware(request, response, next) {
-
-  const creds = utils.getCredentials(request);
-
-  try {
-    checkCredentials(creds);
-    request.userCredentials = creds;
-    next();
-  } catch (e) {
-    response.status(HttpStatus.UNAUTHORIZED).end();
-    return;
-  }
-}
-
-function requireBodyMiddleware(request, response, next) {
-
-  if (typeof request.body !== 'string') {
-    response.status(HttpStatus.UNSUPPORTED_MEDIA_TYPE).end();
-  } else {
-    next();
-  }
-}
-
-function parseRecordMiddleware(request, response, next) {
-
-  try {
-    const record = utils.parseRecord(request.body);
-    request.record = record;
-    next();
-  } catch (e) {
-    response.status(HttpStatus.BAD_REQUEST).type('application/json').send({
-      error: {
-        message: 'Invalid record data'
-      }
-    });
-  }    
-}
+const Middlewares = require('../middlewares');
 
 module.exports = (authorityService, logger) => {
 
@@ -67,7 +22,7 @@ module.exports = (authorityService, logger) => {
       });
   });
 
-  router.put('/:id', requireCredentialsMiddleware, requireBodyMiddleware, parseRecordMiddleware, (request, response) => {
+  router.put('/:id', Middlewares.requireCredentials, Middlewares.requireBody, Middlewares.parseRecord, (request, response) => {
     const recordId = request.params.id;
     const creds = request.userCredentials;
     const record = request.record;
@@ -82,7 +37,7 @@ module.exports = (authorityService, logger) => {
     });  
   });
 
-  router.post('/', requireCredentialsMiddleware, requireBodyMiddleware, parseRecordMiddleware, (request, response) => {
+  router.post('/', Middlewares.requireCredentials, Middlewares.requireBody, Middlewares.parseRecord, (request, response) => {
     const recordId = request.params.id;
     const creds = request.userCredentials;
     const record = request.record;
